Add(minjae/boardDetail): boardDetail Top 컴포넌트 분리 #70
